Add the ability to push when done
If commit_finalize is true, Reissue will commit the changes.
Use push_finalize to configure whether or not these changes should be automatically pushed
to the remote repository.

Set push_finalize to ':branch' to create a separate branch. This is the default.
Or set it to true to push directly to the working branch.
Set it to false to disable this feature independently of commit_finalize.