Fix duplicate finalize version bug
When finalizing the changelog the code was using update which would unshift a new version with the changes.
That lead to duplicate entries of the same version with incorrect values for the release date.
This change does some manual adjustments to the contents of the changelog updater object but fixes
the bug so that we will properly compose unique changes and merge values manually enteried into the file.

Fixed: Duplicate changelog entries with incorrect release dates.