Removed hardcoding and simplified inputs
In the quest to make this more reusable, there were still some more
parts that needed to be updated. The gem name was hard coded for the
version.rb, which may or may not be present for a gem. So tried to work
through other parts which may need more flexibility with a shared flow,
delegating to rake task, removed some input naming and only create a PR
when it is needed.